# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  context.Response:
    properties:
      data: {}
      err:
        type: string
      msg:
        type: string
      success:
        type: boolean
      trace_id:
        type: string
    type: object
  model.AskSession:
    properties:
      bot:
        type: boolean
      content:
        type: string
      created_at:
        type: integer
      id:
        type: integer
      summary:
        type: boolean
      updated_at:
        type: integer
      user_id:
        type: integer
      uuid:
        type: string
    type: object
...
  model.CommentLikeState:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - CommentLikeStateLike
    - CommentLikeStateDislike
...
  model.DiscussionComment:
    properties:
      accepted:
        type: boolean
      bot:
        type: boolean
      content:
        type: string
      created_at:
        type: integer
      dislike:
        type: integer
      id:
        type: integer
      like:
        type: integer
      replies:
        items:
          $ref: '#/definitions/model.DiscussionReply'
        type: array
      updated_at:
        type: integer
      user_avatar:
        type: string
      user_id:
        type: integer
      user_like_state:
        $ref: '#/definitions/model.CommentLikeState'
      user_name:
        type: string
      user_role:
        $ref: '#/definitions/model.UserRole'
    type: object
  model.DiscussionDetail:
    properties:
      alert:
        type: boolean
      associate:
        $ref: '#/definitions/model.DiscussionListItem'
      associate_id:
        type: integer
      bot_unknown:
        type: boolean
      comment:
        type: integer
      comments:
        items:
          $ref: '#/definitions/model.DiscussionComment'
        type: array
      content:
        type: string
      created_at:
        type: integer
      current_user_id:
        type: integer
      dislike:
        type: integer
      forum_id:
        type: integer
      group_ids:
        items:
          type: integer
        type: array
      groups:
        items:
          $ref: '#/definitions/model.DiscussionGroup'
        type: array
      hot:
        type: integer
      id:
        type: integer
      last_visited:
        description: 发帖人上次访问时间
        type: integer
      like:
        type: integer
      members:
        items:
          type: integer
        type: array
      rag_id:
        type: string
      resolved:
        $ref: '#/definitions/model.DiscussionState'
      resolved_at:
        type: integer
      summary:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      tags:
        description: Deprecated
        items:
          type: string
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/model.DiscussionType'
      updated_at:
        type: integer
      user_avatar:
        type: string
      user_id:
        type: integer
      user_like:
        type: boolean
      user_name:
        type: string
      user_role:
        $ref: '#/definitions/model.UserRole'
      uuid:
        type: string
      view:
        type: integer
      visit:
        description: 发帖人访问次数
        type: integer
    type: object
  model.DiscussionGroup:
    properties:
      group_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  model.DiscussionListItem:
    properties:
      associate_id:
        type: integer
      bot_unknown:
        type: boolean
      comment:
        type: integer
      content:
        type: string
      created_at:
        type: integer
      dislike:
        type: integer
      forum_id:
        type: integer
      group_ids:
        items:
          type: integer
        type: array
      hot:
        type: integer
      id:
        type: integer
      last_visited:
        description: 发帖人上次访问时间
        type: integer
      like:
        type: integer
      members:
        items:
          type: integer
        type: array
      rag_id:
        type: string
      resolved:
        $ref: '#/definitions/model.DiscussionState'
      resolved_at:
        type: integer
      similarity:
        type: number
      summary:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      tags:
        description: Deprecated
        items:
          type: string
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/model.DiscussionType'
      updated_at:
        type: integer
      user_avatar:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
      uuid:
        type: string
      view:
        type: integer
      visit:
        description: 发帖人访问次数
        type: integer
    type: object
  model.DiscussionReply:
    properties:
      accepted:
        type: boolean
      bot:
        type: boolean
      content:
        type: string
      created_at:
        type: integer
      dislike:
        type: integer
      id:
        type: integer
      like:
        type: integer
      updated_at:
        type: integer
      user_avatar:
        type: string
      user_id:
        type: integer
      user_like_state:
        $ref: '#/definitions/model.CommentLikeState'
      user_name:
        type: string
      user_role:
        $ref: '#/definitions/model.UserRole'
    type: object
  model.DiscussionState:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - DiscussionStateUnknown
    - DiscussionStateNone
    - DiscussionStateResolved
    - DiscussionStateClosed
    - DiscussionStateInProgress
...
  model.DiscussionType:
    enum:
    - qa
    - feedback
    - blog
    - issue
    type: string
    x-enum-varnames:
    - DiscussionTypeQA
    - DiscussionTypeFeedback
    - DiscussionTypeBlog
    - DiscussionTypeIssue
  model.DocStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    type: integer
    x-enum-varnames:
    - DocStatusUnknown
    - DocStatusApplySuccess
    - DocStatusPendingReview
    - DocStatusPendingApply
    - DocStatusApplyFailed
    - DocStatusAppling
    - DocStatusPendingExport
    - DocStatusExportSuccess
    - DocStatusExportFailed
...
  model.FileType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    - 15
    - 16
    - 17
    - 18
    type: integer
    x-enum-comments:
      FileTypeFile: 未知文件类型
      FileTypeFolder: 文件夹
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - 文件夹
    - 未知文件类型
    - ""
    x-enum-varnames:
    - FileTypeUnknown
    - FileTypeMarkdown
    - FileTypeHTML
    - FileTypeJSON
    - FileTypeURL
    - FileTypeDOCX
    - FileTypeDOC
    - FileTypePPTX
    - FileTypeXLSX
    - FileTypeXLS
    - FileTypePDF
    - FileTypeImage
    - FileTypeCSV
    - FileTypeXML
    - FileTypeZIP
    - FileTypeEPub
    - FileTypeFolder
    - FileTypeFile
    - FileTypeMax
...
  model.ListRes:
    properties:
      total:
        type: integer
    type: object
...
  model.SystemDiscussion:
    properties:
      auto_close:
        type: integer
      content_placeholder:
        type: string
    type: object
...
  model.SystemWebPlugin:
    properties:
      display:
        type: boolean
      enabled:
        type: boolean
    type: object
...
  model.UserRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - UserRoleUnknown
    - UserRoleAdmin
    - UserRoleOperator
    - UserRoleUser
    - UserRoleGuest
    - UserRoleMax
...
  svc.AIInsightDiscussionItem:
    properties:
      created_at:
        type: integer
      deleted:
        type: boolean
      discussion_id:
        type: string
      id:
        type: integer
      rank_id:
        type: integer
      title:
        type: string
      updated_at:
        type: integer
    type: object
...
  svc.DiscussionCreateReq:
    properties:
      content:
        type: string
      forum_id:
        type: integer
      group_ids:
        items:
          type: integer
        type: array
      summary:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/model.DiscussionType'
    required:
    - title
    type: object
...
  svc.DiscussionUpdateReq:
    properties:
      content:
        type: string
      group_ids:
        items:
          type: integer
        type: array
      summary:
        type: string
      title:
        type: string
    type: object
...
  svc.ListWebItem:
    properties:
      created_at:
        type: integer
      desc:
        type: string
      file_type:
        $ref: '#/definitions/model.FileType'
      group_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      status:
        $ref: '#/definitions/model.DocStatus'
      title:
        type: string
      updated_at:
        type: integer
    type: object
...
  svc.StatDiscussionItem:
    properties:
      count:
        type: integer
      key:
        $ref: '#/definitions/model.DiscussionType'
    type: object
  svc.StatDiscussionRes:
    properties:
      accept:
        type: integer
      bot_accept:
        type: integer
      bot_unknown:
        type: integer
      human_resp_time:
        type: integer
    type: object
...
  svc.UpdateWebNotifyReq:
    properties:
      enable:
        type: boolean
    type: object
...
info:
  contact: {}
  description: KoalaQA API swagger document.
  title: KoalaQA API
  version: "2.0"
paths:
...
  /admin/discussion:
    get:
      description: backend list discussions
      parameters:
      - in: query
        name: forum_id
        required: true
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.DiscussionListItem'
                        type: array
                    type: object
              type: object
      summary: backend list discussions
      tags:
      - discussion
  /admin/discussion/ask:
    get:
      description: backend ask session group
      parameters:
      - in: query
        name: content
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        minimum: 1
        name: size
        type: integer
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.AskSession'
                        type: array
                    type: object
              type: object
      summary: backend ask session group
      tags:
      - discussion
  /admin/discussion/ask/session:
    get:
      description: backend ask session
      parameters:
      - in: query
        name: session_id
        required: true
        type: string
      - in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.AskSession'
                        type: array
                    type: object
              type: object
      summary: backend ask session
      tags:
      - discussion
...
  /admin/kb/{kb_id}/web:
    get:
      parameters:
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        minimum: 1
        name: size
        type: integer
      - in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/svc.ListWebItem'
                        type: array
                    type: object
              type: object
      summary: list kb web
      tags:
      - web
  /admin/kb/{kb_id}/web/{doc_id}:
    delete:
      parameters:
      - description: kb_id
        in: path
        name: kb_id
        required: true
        type: integer
      - description: doc_id
        in: path
        name: doc_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: delete kb web
      tags:
      - web
    put:
      parameters:
      - description: kb_id
        in: path
        name: kb_id
        required: true
        type: integer
      - description: doc_id
        in: path
        name: doc_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: update kb web
      tags:
      - web
...
  /admin/rank/ai_insight/{ai_insight_id}/discussion:
    get:
      parameters:
      - description: ai_insight_id
        in: path
        name: ai_insight_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/svc.AIInsightDiscussionItem'
                        type: array
                    type: object
              type: object
      summary: list ai insight discussion
      tags:
      - rank
  /admin/stat/discussion:
    get:
      parameters:
      - in: query
        name: begin
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/svc.StatDiscussionRes'
                  - properties:
                      discussions:
                        items:
                          $ref: '#/definitions/svc.StatDiscussionItem'
                        type: array
                    type: object
              type: object
      summary: stat discussion
      tags:
      - stat
...
  /admin/system/discussion:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SystemDiscussion'
              type: object
      summary: system discussion detail
      tags:
      - system_discussion
    put:
      consumes:
      - application/json
      parameters:
      - description: request params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.SystemDiscussion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: update system discussion config
      tags:
      - system_discussion
...
  /admin/system/web_plugin:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SystemWebPlugin'
              type: object
      summary: web plugin detail
      tags:
      - web_plugin
    put:
      consumes:
      - application/json
      parameters:
      - description: request params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.SystemWebPlugin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: update web plugin config
      tags:
      - web_plugin
...
  /discussion:
    get:
      description: list discussions
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: discussion_ids
        type: array
      - enum:
        - hot
        - new
        - publish
        in: query
        name: filter
        type: string
        x-enum-varnames:
        - DiscussionListFilterHot
        - DiscussionListFilterNew
        - DiscussionListFilterPublish
      - in: query
        name: forum_id
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: group_ids
        type: array
      - in: query
        name: keyword
        type: string
      - in: query
        name: only_mine
        type: boolean
      - in: query
        minimum: 1
        name: page
        type: integer
      - enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        in: query
        name: resolved
        type: integer
        x-enum-varnames:
        - DiscussionStateUnknown
        - DiscussionStateNone
        - DiscussionStateResolved
        - DiscussionStateClosed
        - DiscussionStateInProgress
      - in: query
        minimum: 1
        name: size
        type: integer
      - in: query
        name: stat
        type: boolean
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: tag_ids
        type: array
      - enum:
        - qa
        - feedback
        - blog
        - issue
        in: query
        name: type
        type: string
        x-enum-varnames:
        - DiscussionTypeQA
        - DiscussionTypeFeedback
        - DiscussionTypeBlog
        - DiscussionTypeIssue
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.DiscussionListItem'
                        type: array
                    type: object
              type: object
      summary: list discussions
      tags:
      - discussion
    post:
      consumes:
      - application/json
      description: create discussion
      parameters:
      - description: discussion
        in: body
        name: discussion
        required: true
        schema:
          $ref: '#/definitions/svc.DiscussionCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: create discussion
      tags:
      - discussion
  /discussion/{disc_id}:
    delete:
      consumes:
      - application/json
      description: delete discussion
      parameters:
      - description: disc_id
        in: path
        name: disc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data: {}
              type: object
      summary: delete discussion
      tags:
      - discussion
    get:
      description: detail discussion
      parameters:
      - in: query
        name: no_view
        type: boolean
      - description: disc_id
        in: path
        name: disc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DiscussionDetail'
              type: object
      summary: detail discussion
      tags:
      - discussion
    put:
      consumes:
      - application/json
      description: update discussion
      parameters:
      - description: disc_id
        in: path
        name: disc_id
        required: true
        type: string
      - description: discussion
        in: body
        name: discussion
        required: true
        schema:
          $ref: '#/definitions/svc.DiscussionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data: {}
              type: object
      summary: update discussion
      tags:
      - discussion
...
  /discussion/ask:
    post:
      description: user ask
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: group_ids
        type: array
      - in: query
        name: question
        required: true
        type: string
      - in: query
        name: session_id
        required: true
        type: string
      produces:
      - text/event-stream
      responses: {}
      summary: user ask
      tags:
      - discussion
...
  /user/notify/web:
    post:
      consumes:
      - application/json
      parameters:
      - description: req params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/svc.UpdateWebNotifyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: update notify web switch
      tags:
      - user
...
swagger: "2.0"